broad refactoring, generic clean-up, SQL dummy entry, lots of changes
