+++ src/main/java/com/boundlessgeo/ps/turbopsapi/model/ProjMgmt/ChangeOrderType.java
@@ -1,7 +1,7 @@
 /**
  *
  */
-package com.boundlessgeo.ps.turbopsapi.model;
+package com.boundlessgeo.ps.turbopsapi.model.ProjMgmt;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
@@ -16,11 +16,9 @@ import lombok.ToString;
 @AllArgsConstructor
 @ToString
 @JsonFormat(shape = JsonFormat.Shape.OBJECT)
-public enum ConfigDocType {
-	WORD("MS Word"),
-	EXCEL("MS Excel"),
-	PDF("PDF");
+public enum ChangeOrderType {
+	SCOPE("Scope"), COST("Cost"), POP("PoP");
 
 	@Getter
-	private final String docType;
+	private final String changeOrderType;
 }
