+++ src/main/java/com/boundlessgeo/ps/turbopsapi/model/ProjMgmt/ConfigDoc.java
@@ -1,13 +1,16 @@
 /**
  *
  */
-package com.boundlessgeo.ps.turbopsapi.model;
+package com.boundlessgeo.ps.turbopsapi.model.ProjMgmt;
 
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
+import com.boundlessgeo.ps.turbopsapi.model.AuditableObject;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import lombok.Getter;
@@ -21,12 +24,30 @@ import lombok.ToString;
 @SuppressWarnings("serial")
 @Entity
 @ToString(includeFieldNames = true, exclude = "project")
-public class ProjectNote extends AuditableObject {
+public class ConfigDoc extends AuditableObject {
 	@Getter
 	@Setter
-	private String note;
+	private String docName;
 
-	// TODO: Need to check ProjectNote JSON serialization
+	@Getter
+	@Setter
+	@Enumerated(EnumType.STRING)
+	private ConfigDocType docType;
+
+	@Getter
+	@Setter
+	private long version;
+
+	// TODO: Probably should be replaced by an object representing "user".
+	@Getter
+	@Setter
+	private String owner;
+
+	@Getter
+	@Setter
+	private String location;
+
+	// TODO: Need to check ConfigDoc JSON serialization
 	// Seems like both the exclude in the @ToString above and the @JsonIgnore
 	// below are needed for correct JSON serialization, otherwise stack
 	// overflow errors will occur.
