+++ src/main/java/com/boundlessgeo/ps/turbopsapi/model/ProjMgmt/Contingency.java
@@ -1,15 +1,14 @@
 /**
  *
  */
-package com.boundlessgeo.ps.turbopsapi.model;
+package com.boundlessgeo.ps.turbopsapi.model.ProjMgmt;
 
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
+import com.boundlessgeo.ps.turbopsapi.model.AuditableObject;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import lombok.Getter;
@@ -18,42 +17,28 @@ import lombok.ToString;
 
 /**
  * @author ssengupta
- *
  */
 @SuppressWarnings("serial")
 @Entity
-@ToString(includeFieldNames = true, exclude = "project")
-public class ConfigDoc extends AuditableObject {
-	@Getter
-	@Setter
-	private String docName;
+@ToString(includeFieldNames = true, exclude = "risk")
+public class Contingency extends AuditableObject {
 
 	@Getter
 	@Setter
-	@Enumerated(EnumType.STRING)
-	private ConfigDocType docType;
+	private String trigger;
 
 	@Getter
 	@Setter
-	private long version;
+	private String description;
 
-	// TODO: Probably should be replaced by an object representing "user".
-	@Getter
-	@Setter
-	private String owner;
-
-	@Getter
-	@Setter
-	private String location;
-
-	// TODO: Need to check ConfigDoc JSON serialization
+	// TODO: Need to check Contingency JSON serialization
 	// Seems like both the exclude in the @ToString above and the @JsonIgnore
 	// below are needed for correct JSON serialization, otherwise stack
 	// overflow errors will occur.
 	@Getter
 	@Setter
 	@ManyToOne(fetch = FetchType.EAGER)
-	@JoinColumn(name = "PROJECT_ID")
+	@JoinColumn(name = "RISK_ID")
 	@JsonIgnore
-	private Project project;
+	private Risk risk;
 }
