+++ src/main/java/com/boundlessgeo/ps/turbopsapi/model/ProjMgmt/ProjectManager.java
@@ -1,13 +1,15 @@
 /**
  *
  */
-package com.boundlessgeo.ps.turbopsapi.model;
+package com.boundlessgeo.ps.turbopsapi.model.ProjMgmt;
 
 import java.util.List;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.OneToMany;
+
+import com.boundlessgeo.ps.turbopsapi.model.AuditableObject;
 
 import lombok.Getter;
 import lombok.Setter;
