+++ src/main/java/com/boundlessgeo/ps/turbopsapi/model/ProjMgmt/ProjectNote.java
@@ -1,13 +1,14 @@
 /**
  *
  */
-package com.boundlessgeo.ps.turbopsapi.model;
+package com.boundlessgeo.ps.turbopsapi.model.ProjMgmt;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
+import com.boundlessgeo.ps.turbopsapi.model.AuditableObject;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import lombok.Getter;
@@ -16,28 +17,24 @@ import lombok.ToString;
 
 /**
  * @author ssengupta
+ *
  */
 @SuppressWarnings("serial")
 @Entity
-@ToString(includeFieldNames = true, exclude = "risk")
-public class Contingency extends AuditableObject {
-
+@ToString(includeFieldNames = true, exclude = "project")
+public class ProjectNote extends AuditableObject {
 	@Getter
 	@Setter
-	private String trigger;
+	private String note;
 
-	@Getter
-	@Setter
-	private String description;
-
-	// TODO: Need to check Contingency JSON serialization
+	// TODO: Need to check ProjectNote JSON serialization
 	// Seems like both the exclude in the @ToString above and the @JsonIgnore
 	// below are needed for correct JSON serialization, otherwise stack
 	// overflow errors will occur.
 	@Getter
 	@Setter
 	@ManyToOne(fetch = FetchType.EAGER)
-	@JoinColumn(name = "RISK_ID")
+	@JoinColumn(name = "PROJECT_ID")
 	@JsonIgnore
-	private Risk risk;
+	private Project project;
 }
