+++ src/main/java/com/boundlessgeo/ps/turbopsapi/model/ProjMgmt/Risk.java
@@ -1,15 +1,20 @@
 /**
  *
  */
-package com.boundlessgeo.ps.turbopsapi.model;
+package com.boundlessgeo.ps.turbopsapi.model.ProjMgmt;
 
+import java.util.List;
+
+import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
 
+import com.boundlessgeo.ps.turbopsapi.model.AuditableObject;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import lombok.Getter;
@@ -23,25 +28,40 @@ import lombok.ToString;
 @SuppressWarnings("serial")
 @Entity
 @ToString(includeFieldNames = true, exclude = "project")
-public class ChangeOrder extends AuditableObject {
-	@Getter
-	@Setter
-	private String initiatedBy;
+public class Risk extends AuditableObject {
 
 	@Getter
 	@Setter
-	private String reason;
+	private String riskName;
 
 	@Getter
 	@Setter
 	@Enumerated(EnumType.STRING)
-	private ChangeOrderType changeOrderType;
+	private RiskType riskType;
 
 	@Getter
 	@Setter
-	private long cost;
+	@Enumerated(EnumType.STRING)
+	private RiskStatus riskStatus;
 
-	// TODO: Need to check ChangeOrder JSON serialization
+	@Getter
+	@Setter
+	private Double probability;
+
+	@Getter
+	@Setter
+	private String impact;
+
+	@Getter
+	@Setter
+	private String rating;
+
+	// TODO: Probably should be replaced by an object representing "user".
+	@Getter
+	@Setter
+	private String owner;
+
+	// TODO: Need to check Risk JSON serialization
 	// Seems like both the exclude in the @ToString above and the @JsonIgnore
 	// below are needed for correct JSON serialization, otherwise stack
 	// overflow errors will occur.
@@ -51,4 +71,9 @@ public class ChangeOrder extends AuditableObject {
 	@JoinColumn(name = "PROJECT_ID")
 	@JsonIgnore
 	private Project project;
+
+	@Getter
+	@Setter
+	@OneToMany(mappedBy = "risk", cascade = CascadeType.ALL)
+	private List<Contingency> contingencies;
 }
