+++ src/main/java/com/boundlessgeo/ps/turbopsapi/model/ProjMgmt/RiskType.java
@@ -1,7 +1,7 @@
 /**
  *
  */
-package com.boundlessgeo.ps.turbopsapi.model;
+package com.boundlessgeo.ps.turbopsapi.model.ProjMgmt;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
@@ -16,12 +16,12 @@ import lombok.ToString;
 @AllArgsConstructor
 @ToString
 @JsonFormat(shape = JsonFormat.Shape.OBJECT)
-public enum RiskStatus {
-	IMMINENT("Imminent"),
-	ONGOING("Ongoing"),
-	MITIGATED("Mitigated"),
-	FAILED("Failed");
+public enum RiskType {
+	CUSTSAT("Cust Sat"),
+	COSTOVERRUN("Cost Overrun"),
+	RESOURCEUNAVAILABILITY("Resource Unavailability"),
+	POPOVERRUN("PoP Overrun");
 
 	@Getter
-	private String status;
+	private final String type;
 }
