+++ src/main/java/com/boundlessgeo/ps/turbopsapi/repositories/ProjMgmt/ProjectRepository.java
@@ -1,7 +1,7 @@
 /**
  *
  */
-package com.boundlessgeo.ps.turbopsapi.repositories;
+package com.boundlessgeo.ps.turbopsapi.repositories.ProjMgmt;
 
 import java.util.List;
 
@@ -9,7 +9,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
-import com.boundlessgeo.ps.turbopsapi.model.Project;
+import com.boundlessgeo.ps.turbopsapi.model.ProjMgmt.Project;
 
 /**
  * @author ssengupta
